refactor: extract model enums to models/enums.py (#652)
Move all StrEnum classes from the monolithic models.py into a dedicated
models/enums.py module. The models/ package __init__.py re-exports every
enum so existing imports remain valid.
